+++ rpc/src/v1/types/hash.rs
@@ -20,7 +20,7 @@ use std::cmp::Ordering;
 use std::hash::{Hash, Hasher};
 use serde;
 use rustc_serialize::hex::{ToHex, FromHex};
-use util::{H64 as Eth64, H256 as EthH256, H520 as EthH520, H2048 as Eth2048, Address};
+use util::{H64 as Eth64, H256 as EthH256, H520 as EthH520, H2048 as Eth2048, H160 as Eth160};
 
 macro_rules! impl_hash {
 	($name: ident, $other: ident, $size: expr) => {
@@ -143,7 +143,7 @@ macro_rules! impl_hash {
 }
 
 impl_hash!(H64, Eth64, 8);
-impl_hash!(H160, Address, 20);
+impl_hash!(H160, Eth160, 20);
 impl_hash!(H256, EthH256, 32);
 impl_hash!(H520, EthH520, 65);
 impl_hash!(H2048, Eth2048, 256);
